backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
crewai>=0.80.0,<1.0.0
crewai-tools>=0.35.0,<1.0.0
langchain-google-genai==2.1.2
langchain==0.3.14
python-dotenv==1.0.1
pydantic==2.10.4
sse-starlette==2.2.1

# RAG dependencies
pymupdf==1.24.11
chromadb==0.5.23
python-multipart==0.0.17
